Refactor camera ray generation and add model loading
- Changed the pixel coordinate type in Camera::generateRay from glm::vec2 to glm::ivec2 for better integer handling.
- Introduced a new Model class to handle loading 3D models from OBJ files, including parsing vertices and normals.
- Added Shape interface with Sphere and Triangle implementations for intersection tests.
- Updated Ray structure to include a method for calculating points along the ray.

+++ include/camera.h
@@ -13,7 +13,7 @@ class Camera
 public:
     Camera(Image& image, const glm::vec3& pos, const glm::vec3& viewpoint, float fovy);
 
-    Ray generateRay(const glm::vec2& pixelCoord, const glm::vec2& offset = {0.5f, 0.5f}) const;
+    Ray generateRay(const glm::ivec2& pixelCoord, const glm::vec2& offset = {0.5f, 0.5f}) const;
 
 private:
     Image&    mImage;
